Virtual Power Plant and VPP Operator added
This commit introduces the class for the „Virtual Power Plant“ that contains all the components, that should be simulated.
Furthermore the class „VPPOperator“ was added. It provides basic logic for operating a „VirtualPowerPlant“-object. It needs to be subclassed to accomodate for the different stategies of operation (e. g. machine learning and forecasting).