Implemented VPPEnvironment and VPPUseCase
The VPPEnvironment class provides information about the environment the virtual power plant is operated in. This contains, for example, the weather, location and external regulary influences.

The VPPUseCase class is considered to represent different usage of components by different people.

Both VPPEnvironment and VPPUseCase are passed in as a parameter of a VPPComponent and therefore can be accessed for creating the simulated profiles.